+++ tasks/shared/models/index.ts
@@ -1,3 +1,4 @@
-export * from './connection';
-export * from './pullRequest';
-export * from './release';
+export * from './connection'
+export * from './pullRequest'
+export * from './release'
+export * from './variables'