+++ tasks/shared/models/variables.ts
@@ -1,0 +1,11 @@
+export interface TaskVariables {
+  apiKey: string
+  apiToken: string
+  email: string
+  account: string
+  devBranch: string
+  forcePublish: boolean
+  deploy: boolean
+  beta: boolean
+  mergeIntoDevelop: boolean
+}